requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
Authlib==1.2.0
Babel==2.10.3
beautifulsoup4==4.11.2
beniget==0.4.1
blinker==1.5
Bottleneck==1.3.5
Brlapi==0.8.4
Brotli==1.0.9
certifi==2022.9.24
chardet==5.1.0
click==8.1.3
cloud-init==23.2.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.8.0
distro-info==1.5
et-xmlfile==1.0.1
exceptiongroup==1.1.0
Flask==2.2.2
flask_oidc
fonttools==4.38.0
fs==2.4.16
gast==0.5.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
jaraco.classes==3.2.1
jdcal==1.0
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.6.0
keycloak
keyring==23.9.3
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
llvmlite==0.39.1
louis==3.24.0
lxml==4.9.2
lz4==4.0.2+dfsg
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.5.2
mdurl==0.1.2
more-itertools==8.10.0
mpmath==0.0.0
mysql-connector-python==8.0.15
netifaces==0.11.0
numba==0.56.4
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
outcome==1.2.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
pexpect==4.8.0
Pillow==9.4.0
pluggy==1.0.0+repack
ply==3.11
protobuf==4.21.12
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.14.0
PyGObject==3.44.1
pyinotify==0.9.6
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyserial==3.5
pytest==7.2.1
python-apt==2.5.3+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.0
pythran==0.11.0
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
rich==13.3.1
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.3
selenium==4.8.2
simplejson==3.18.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.4
stripe
sympy==1.11.1
systemd-python==235
tables==3.7.0
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.14.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.12
wadllib==1.3.6
webdriver-manager==4.0.0
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
